@@ -20,32 +20,39 @@ dependencies:
   intl: ^0.20.2
   flutter_drawing_board: ^0.9.8
   firebase_core: ^3.12.1
-  cloud_firestore: ^5.6.5
-  firebase_auth: ^5.5.1
+  cloud_firestore: ^5.6.7   # Güncel sürüm
+  firebase_auth: ^5.5.3      # Güncel sürüm
   cached_network_image: ^3.4.1
   js: ^0.7.2
-  url_launcher: ^6.3.1
-
+  url_launcher: ^6.3.1    # Güncel sürüm
+  
   youtube_player_flutter: ^9.1.1
-  camera: ^0.10.5+9
-  tflite_flutter: ^0.10.0  # Burada eklenen tflite_flutter paketi
+  camera: ^0.11.1         # Güncel sürüm
+  tflite_flutter: ^0.11.0  # Güncel sürüm
   image_picker: ^1.1.2
-  google_fonts: ^6.2.1 
+  google_fonts: ^6.2.1
 
   shelf: ^1.4.2
   shelf_router: ^1.1.4
-  http: ^1.3.0
+  http: ^1.4.0           # Güncel sürüm
   flutter_dotenv: ^5.2.1
 
-
 dev_dependencies:
+  flutter_launcher_icons: ^0.13.1  # En güncel versiyonu kontrol edebilirsin
+
+flutter_icons:
+  android: true
+  ios: true
+  image_path: "assets/icon/app_icon.png"
+  
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0   # Güncel sürüm
 
 flutter:
   uses-material-design: true
   assets:
+    - assets/logo.jpg
     - assets/images/domates.jpg
     - assets/images/gül.jpg
     - assets/images/lale.jpg
@@ -53,3 +60,4 @@ flutter:
     - assets/images/patlıcan.jpg
     - assets/images/yeşil.jpg
     - app.env
+
